% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TinglanToolsmywrappers.R
\name{dplyrWrapper}
\alias{dplyrWrapper}
\title{dplyr wrapper function for acquiring marginals using function dplyr}
\usage{
dplyrWrapper(x)
}
\arguments{
\item{x}{data-frame}
}
\description{
manipulate margins and select data
}
