% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TinglanToolsRfunctions.R
\name{newA}
\alias{newA}
\title{HW problem to standaridize columns}
\usage{
newA(x)
}
\arguments{
\item{a}{data.frame}
}
\value{
list
}
\description{
Computes (x−mean(x))/sd(x) for each column
}
\examples{
load(url("http://www.stat.umn.edu/geyer/3701/data/q2p3.rda"))
newA(a)
}
