% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TinglanTools-package.R
\docType{package}
\name{TinglanTools-package}
\alias{TinglanTools}
\alias{TinglanTools-package}
\title{TinglanTools: My Tools for STAT 3701}
\description{
This are several functions for STAT 3701.
}
\author{
\strong{Maintainer}: Tinglan Li \email{lixx5213@umn.edu}

}
\keyword{internal}
